hw-2

+++ app/src/main/java/com/example/shmryandex/domain/entity/Account.kt
@@ -3,6 +3,6 @@ package com.example.shmryandex.domain.entity
 data class Account(
     val id: Int,
     val name: String,
-    val balance: Int,
+    val balance: Double,
     val currency: String
 )
